Rename room package to db
