+++ app/src/main/kotlin/org/bspb/smartbirds/pro/db/Converters.kt
@@ -1,4 +1,4 @@
-package org.bspb.smartbirds.pro.room
+package org.bspb.smartbirds.pro.db
 
 import androidx.room.TypeConverter
