+++ app/src/main/kotlin/org/bspb/smartbirds/pro/db/dao/NomenclatureDao.kt
@@ -1,10 +1,10 @@
-package org.bspb.smartbirds.pro.room.dao
+package org.bspb.smartbirds.pro.db.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
-import org.bspb.smartbirds.pro.room.model.NomenclatureModel
+import org.bspb.smartbirds.pro.db.model.NomenclatureModel
 
 @Dao
 abstract class NomenclatureDao {
