+++ app/src/main/kotlin/org/bspb/smartbirds/pro/db/dao/NomenclatureUsesCountDao.kt
@@ -1,10 +1,10 @@
-package org.bspb.smartbirds.pro.room.dao
+package org.bspb.smartbirds.pro.db.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Update
-import org.bspb.smartbirds.pro.room.model.NomenclatureUsesCount
+import org.bspb.smartbirds.pro.db.model.NomenclatureUsesCount
 
 @Dao
 abstract class NomenclatureUsesCountDao {
