+++ app/src/main/kotlin/org/bspb/smartbirds/pro/db/dao/ZoneDao.kt
@@ -1,11 +1,11 @@
-package org.bspb.smartbirds.pro.room.dao
+package org.bspb.smartbirds.pro.db.dao
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
-import org.bspb.smartbirds.pro.room.model.ZoneModel
+import org.bspb.smartbirds.pro.db.model.ZoneModel
 
 @Dao
 abstract class ZoneDao {
