+++ app/src/main/kotlin/org/bspb/smartbirds/pro/sync/ZonesManager.kt
@@ -8,8 +8,8 @@ import org.androidannotations.annotations.Bean
 import org.androidannotations.annotations.EBean
 import org.androidannotations.annotations.RootContext
 import org.bspb.smartbirds.pro.backend.Backend
-import org.bspb.smartbirds.pro.room.SmartBirdsDatabase
-import org.bspb.smartbirds.pro.room.model.ZoneModel
+import org.bspb.smartbirds.pro.db.SmartBirdsDatabase
+import org.bspb.smartbirds.pro.db.model.ZoneModel
 import org.bspb.smartbirds.pro.tools.Reporting
 import org.bspb.smartbirds.pro.tools.SBGsonParser
 import java.io.IOException
