+++ app/src/main/kotlin/org/bspb/smartbirds/pro/utils/MonitoringManager.kt
@@ -15,9 +15,9 @@ import org.bspb.smartbirds.pro.content.TrackingLocation
 import org.bspb.smartbirds.pro.enums.EntryType
 import org.bspb.smartbirds.pro.repository.FormRepository
 import org.bspb.smartbirds.pro.repository.MonitoringRepository
-import org.bspb.smartbirds.pro.room.model.Form
-import org.bspb.smartbirds.pro.room.model.MonitoringModel
-import org.bspb.smartbirds.pro.room.model.Tracking
+import org.bspb.smartbirds.pro.db.model.Form
+import org.bspb.smartbirds.pro.db.model.MonitoringModel
+import org.bspb.smartbirds.pro.db.model.Tracking
 import org.bspb.smartbirds.pro.tools.SBGsonParser
 import java.text.DateFormat
 import java.text.SimpleDateFormat
